# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Person {
  id: string;
  name: string;
  /** @nullable */
  company: string | null;
  /** @nullable */
  role: string | null;
  /** @nullable */
  location: string | null;
  /** @nullable */
  howMet: string | null;
  /** @nullable */
  dateMet: string | null;
  /** @nullable */
  notes: string | null;
  /** @nullable */
  lastContacted: string | null;
  tags: string[];
  createdAt: string;
}

export interface PersonInput {
  /** @minLength 1 */
  name: string;
  company?: string;
  role?: string;
  location?: string;
  howMet?: string;
  dateMet?: string;
  notes?: string;
  tags?: string[];
}

export interface PersonUpdate {
  /** @minLength 1 */
  name?: string;
  company?: string;
  role?: string;
  location?: string;
  howMet?: string;
  dateMet?: string;
  notes?: string;
  tags?: string[];
}

export interface Interaction {
  id: string;
  personId: string;
  date: string;
  summary: string;
  /** @nullable */
  rawNote: string | null;
}

export interface InteractionInput {
  /** @minLength 1 */
  summary: string;
  rawNote?: string;
  date: string;
}

export interface Connection {
  id: string;
  personAId: string;
  personBId: string;
  relationshipType: string;
  /** @nullable */
  notes: string | null;
}

export type PersonDetail = Person & {
  interactions: Interaction[];
  connections: Connection[];
};

export interface ExtractedRelationship {
  name: string;
  /** @nullable */
  company: string | null;
  /** @nullable */
  role: string | null;
  /** @nullable */
  location: string | null;
  interests: string[];
  connectedTo: string[];
  /** @nullable */
  context: string | null;
  date: string;
  /** @nullable */
  status: string | null;
}

export interface CaptureExtractInput {
  /** @minLength 1 */
  note: string;
}

export interface CaptureExtractResult {
  extracted: ExtractedRelationship;
  isExistingPerson: boolean;
  rawNote: string;
}

export interface CaptureConfirmInput {
  /** @minLength 1 */
  name: string;
  company?: string;
  role?: string;
  location?: string;
  context?: string;
  interests?: string[];
  connectedTo?: string[];
  status?: string;
  date: string;
  /** @minLength 1 */
  rawNote: string;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude?: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude?: number;
}

export interface CaptureConfirmResult {
  person: Person;
  created: boolean;
}

export interface ReconnectPrompt {
  person: Person;
  /** @minimum 0 */
  daysSinceContact: number;
  /** @nullable */
  lastInteraction: string | null;
}

export type GraphNodesItem = {
  id: string;
  name: string;
  /** @nullable */
  company: string | null;
  initials: string;
  tags: string[];
};

export interface Graph {
  nodes: GraphNodesItem[];
  edges: Connection[];
}

export type ListPeopleParams = {
search?: string;
};