# lib/api-zod/src/generated/types/captureConfirmInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaptureConfirmInput {
  /** @minLength 1 */
  name: string;
  company?: string;
  role?: string;
  location?: string;
  context?: string;
  interests?: string[];
  connectedTo?: string[];
  status?: string;
  date: Date;
  /** @minLength 1 */
  rawNote: string;
  /**
     * @minimum -90
     * @maximum 90
     */
  latitude?: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  longitude?: number;
}