# lib/api-zod/src/generated/types/captureConfirmResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Person } from './person';

export interface CaptureConfirmResult {
  person: Person;
  created: boolean;
}