# lib/api-zod/src/generated/types/captureExtractResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedRelationship } from './extractedRelationship';

export interface CaptureExtractResult {
  extracted: ExtractedRelationship;
  isExistingPerson: boolean;
  rawNote: string;
}