# lib/api-zod/src/generated/types/extractedRelationship.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedRelationship {
  name: string;
  /** @nullable */
  company: string | null;
  /** @nullable */
  role: string | null;
  /** @nullable */
  location: string | null;
  interests: string[];
  connectedTo: string[];
  /** @nullable */
  context: string | null;
  date: Date;
  /** @nullable */
  status: string | null;
}